% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{annotate_foci_counting}
\alias{annotate_foci_counting}
\title{annotate_foci_counting}
\usage{
annotate_foci_counting(
  img_file,
  cell_count,
  img_orig,
  img_orig_foci,
  artificial_amp_factor,
  strands,
  coincident_foci,
  foci_label
)
}
\arguments{
\item{img_file}{cell's file name}

\item{cell_count}{unique cell counter}

\item{img_orig}{original strand crop}

\item{img_orig_foci}{cropped foci channel}

\item{artificial_amp_factor}{amplification factor}

\item{strands}{black white mask of strand channel}

\item{coincident_foci}{mask of overlap between strand and foci channel}

\item{foci_label}{black and white mask of foci channel}
}
\description{
Contains all plotting routines for count foci annotation
}
