% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{annotate_foci_counting_adjusted}
\alias{annotate_foci_counting_adjusted}
\title{annotate_foci_counting_adjusted}
\usage{
annotate_foci_counting_adjusted(
  img_file,
  cell_count,
  img_orig,
  img_orig_foci,
  artificial_amp_factor,
  strands,
  coincident_foci,
  foci_label,
  alone_foci,
  percent_px,
  foci_per_cell
)
}
\arguments{
\item{img_file}{cell's file name}

\item{cell_count}{unique cell counter}

\item{img_orig}{original strand crop}

\item{img_orig_foci}{cropped foci channel}

\item{artificial_amp_factor}{amplification factor}

\item{strands}{black white mask of strand channel}

\item{coincident_foci}{mask of overlap between strand and foci channel}

\item{foci_label}{black and white mask of foci channel}

\item{alone_foci}{estimated number of foci that are NOT on a strand.}

\item{percent_px}{percentage of foci mask that coincides with strand channel
small number indicates potentially problematic image.}

\item{foci_per_cell}{number of foci counted per cell}
}
\value{
displays key steps from raw image to coincident foci count
}
\description{
Contains all plotting routines for count foci annotation
}
