% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{append_data_frame}
\alias{append_data_frame}
\title{append_data_frame}
\usage{
append_data_frame(
  WT_str,
  KO_str,
  WT_out,
  KO_out,
  img_file,
  foci_areas,
  df_cells,
  cell_count,
  stage,
  foci_per_cell,
  image_mat,
  percent_px,
  alone_foci
)
}
\arguments{
\item{WT_str}{string in filename corresponding to wildtype genotype.
Defaults to ++.}

\item{KO_str}{string in filename corresponding to knockout genotype.
Defaults to --.}

\item{WT_out}{string in output csv in genotype column, for knockout.
Defaults to +/+.}

\item{KO_out}{string in output csv in genotype column, for knockout.
Defaults to -/-.}

\item{img_file}{cell's file name}

\item{foci_areas}{pixel area of each foci}

\item{df_cells}{current data frame}

\item{cell_count}{unique cell counter}

\item{stage}{meiotic stage}

\item{foci_per_cell}{foci count for cell}

\item{image_mat}{matrix with all pixel values above zero}

\item{percent_px}{percentage of foci mask that coincides with strand channel
small number indicates potentially problematic image.}

\item{alone_foci}{estimated number of foci that are NOT on a strand.}
}
\description{
applies new row to data frame
}
