% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{auto_crop}
\alias{auto_crop}
\title{auto_crop}
\usage{
auto_crop(file_list, img_path, crop_method = "regular")
}
\arguments{
\item{file_list}{The file list}

\item{img_path}{The path}
}
\value{
cropped SC and foci channels around single cells, regardless of stage
}
\description{
crop an image around each viable cell candidate.
}
