% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{count_foci}
\alias{count_foci}
\title{count_foci}
\usage{
count_foci(
  img_path,
  stage = "pachytene",
  offset_px = 0.2,
  offset_factor = 2,
  brush_size = 3,
  brush_sigma = 3,
  foci_norm = 0.01,
  annotate = "off"
)
}
\arguments{
\item{img_path, }{path of crops}

\item{stage, }{description}

\item{offset_px, }{description}

\item{offset_factor, }{description}

\item{brush_size, }{description}

\item{brush_sigma, }{description}

\item{foci_norm, }{description}

\item{annotate, }{description}
}
\value{
foci count per cell
}
\description{
Count coincident foci in cropped SC and foci channel per cell
}
