% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances_general.R
\name{find_start}
\alias{find_start}
\title{find_start}
\usage{
find_start(window, noise_gone, cx, cy)
}
\arguments{
\item{window}{window size that we want to look in}

\item{noise_gone}{Image containing only the SC of interest without background}

\item{cx, cy}{The location of the "centre of intensity" using computeFeatures (not necessarily on SC)}
}
\value{
The location of the starting point (on the SC)
}
\description{
Finds an appropriate place to start counting along an SC by looking in a window around the "centre of intensity"
}
