% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{first_shot_out}
\alias{first_shot_out}
\title{first_shot_out}
\usage{
first_shot_out(
  chosen_dir,
  ix1,
  ix2,
  iy1,
  iy2,
  distance_strand,
  distance_strand_2
)
}
\arguments{
\item{chosen_dir}{The brightest direction of a line passing through starting point}

\item{ix1, }{description}

\item{ix2, }{description}

\item{iy1, }{description}

\item{iy2, }{description}

\item{distance_strand, }{description}

\item{distance_strand_2, }{description}
}
\value{
New sub square for first and second branch
}
\description{
Moves one pixel away from the starting point
}
