% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{get_blobs}
\alias{get_blobs}
\title{get_blobs}
\usage{
get_blobs(img_orig, crop_method = "regular")
}
\arguments{
\item{img_orig}{Original image}
}
\value{
Mask with cell candidates
}
\description{
Makes mask of all objects bright enough
}
