% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{get_C1}
\alias{get_C1}
\title{get_C1}
\usage{
get_C1(foci_areas, foci_per_cell, C_weigh_foci_number)
}
\arguments{
\item{foci_areas}{pixel area of each foci}

\item{foci_per_cell}{foci count for cell}

\item{C_weigh_foci_number}{choose crispness criteria- defaults to TRUE to use
C1 (weighing with number). Otherwise set to FALSE to use C2}
}
\value{
statistic to comapre to crisp_criteria
}
\description{
calculates the statistic to compare to crisp_criteria, which determines
whether the foci count will be reliable
}
