% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(
  strands,
  num_strands,
  new_img,
  foci_label,
  SC_lengths,
  foci_count_strand,
  strand_iter,
  file,
  annotate,
  eccentricity_min,
  max_strand_area
)
}
\arguments{
\item{strands, }{A black white mask with SCs as objects}

\item{num_strands, }{description}

\item{new_img, }{description}

\item{foci_label, }{A black white mask with foci as objects}

\item{SC_lengths, }{description}

\item{foci_count_strand, }{description}

\item{strand_iter, }{description}

\item{file, }{description}

\item{annotate, }{description}

\item{eccentricity_min, }{description}

\item{max_strand_area, }{description}
}
\value{
A list of distances
}
\description{
Creates mask for SC channel
}
