% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{get_distance_between_two}
\alias{get_distance_between_two}
\title{get_distance_between_two}
\usage{
get_distance_between_two(
  distance_strand,
  distance_strand_2,
  per_strand,
  foci_label,
  walkers,
  noise_gone,
  start_x,
  start_y,
  start_x2,
  start_y2,
  start_dir,
  cx,
  cy,
  mean_x,
  mean_y,
  strand_iter,
  file,
  annotation,
  cell_count,
  uid_strand
)
}
\arguments{
\item{distance_strand, }{total distance along first branch}

\item{distance_strand_2, }{total distance along second branch}

\item{per_strand, }{Mask with colocalizing foci}

\item{foci_label, }{black white mask with foci as objects, not necessarily on SC. Needed for computefeatures.}

\item{walkers, }{black white mask containing the line that traces through the middle of the SC. Computed earlier in get_distance}

\item{noise_gone, }{Black white mask with the single SC to be measured}

\item{start_x, }{x pixel location of where first branch terminated}

\item{start_y, }{y pixel location of where first branch terminated}

\item{start_x2, }{x pixel location of where second branch terminated}

\item{start_y2, }{y pixel location of where second branch terminated}

\item{start_dir, }{direction that the first branch was traveling along when it terminated}

\item{cx, }{centre of intensity x of the SC from computefeatures (annotation purposes only)}

\item{cy, }{centre of intensity y of the SC from computefeatures (annotation purposes only)}

\item{mean_x, }{starting point x that the two branches move away from to trace out the SC (annotation purposes only)}

\item{mean_y, }{starting point x that the two branches move away from to trace out the SC (annotation purposes only)}

\item{strand_iter, }{Strand number in iteration over all in cell}

\item{file, }{original filename that cell candidate came from. Used to identify e.g. genotype for data frame.}

\item{annotation, }{Choice to output pipeline choices (recommended to knit)}

\item{cell_count}{Unique cell number}

\item{uid_strand}{Unique strand number}
}
\value{
List of fractional distances between foci for all SCs with two. Optional: total distances of SCs. Optional: images of all resulting traces/ foci locations.
}
\description{
Calculates the pixel distance
}
