% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances_general.R
\name{get_distance_general}
\alias{get_distance_general}
\title{get_distance_general
Creates mask for SC channel}
\usage{
get_distance_general(
  strands,
  num_strands,
  new_img,
  foci_label,
  foci_count_strand,
  strand_iter,
  img_file,
  annotation,
  eccentricity_min,
  max_strand_area,
  cell_count,
  KO_str,
  WT_str,
  KO_out,
  WT_out
)
}
\arguments{
\item{strands, }{A black white mask with SCs as objects}

\item{num_strands, }{Number of individual strands on SC mask}

\item{new_img, }{Original strand/dna/SYCP3 channel image with noise removed.}

\item{foci_label, }{A black white mask with foci as objects}

\item{foci_count_strand, }{Number of foci counted located on the one SC}

\item{strand_iter, }{Strand number in iteration over all in cell}

\item{img_file, }{original filename that cell candidate came from. Used to identify e.g. genotype for data frame.}

\item{annotation, }{Choice to output pipeline choices (recommended to knit)}

\item{eccentricity_min, }{The minimum eccentricity (from computefeatures) of a strand to proceed with measuring}

\item{max_strand_area, }{Maximum pixel area of a strand}

\item{cell_count}{Unique cell counter}

\item{KO_str}{string in filename corresponding to knockout genotype. Defaults to --.}

\item{WT_str}{string in filename corresponding to wildtype genotype. Defaults to ++.}

\item{KO_out}{string in output csv in genotype column, for knockout. Defaults to -/-.}

\item{WT_out}{string in output csv in genotype column, for knockout. Defaults to +/+.}
}
\value{
Data frame with properties of synaptonemal (SC) measurements
}
\description{
get_distance_general
Creates mask for SC channel
}
