% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances_general.R
\name{get_first_dir}
\alias{get_first_dir}
\title{get_first_dir}
\usage{
get_first_dir(noise_gone, ix, iy, window)
}
\arguments{
\item{noise_gone}{Image containing only the SC of interest without background}

\item{ix, iy}{Pixel locations of the starting point (on the SC)}

\item{window}{window size where we compute local gradients}
}
\value{
string: The direction of brightest intensity from starting point
}
\description{
Finds an appropriate place to start counting along an SC by looking in a window around the "centre of intensity"
}
