% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{get_next_first_dir}
\alias{get_next_first_dir}
\title{get_next_first_dir}
\usage{
get_next_first_dir(
  new_square_1,
  ix1,
  iy1,
  dir_1,
  window,
  chosen_dir,
  distance_strand,
  first_dir
)
}
\arguments{
\item{new_square_1, }{description}

\item{ix1, }{description}

\item{iy1, }{description}

\item{dir_1, }{description}

\item{window, }{description}

\item{chosen_dir}{The brightest direction of the previous step}

\item{distance_strand, }{description}

\item{first_dir, }{description}
}
\value{
New sub square for first branch after taking one step
}
\description{
Moves one pixel away one first branch
}
