% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{get_next_second_dir}
\alias{get_next_second_dir}
\title{get_next_second_dir}
\usage{
get_next_second_dir(
  new_square_2,
  ix2,
  iy2,
  dir_2,
  window,
  chosen_dir,
  distance_strand_2,
  second_dir
)
}
\arguments{
\item{new_square_2, }{description}

\item{ix2, }{description}

\item{iy2, }{description}

\item{dir_2, }{description}

\item{window, }{description}

\item{chosen_dir}{The brightest direction of the previous step}

\item{distance_strand_2, }{description}

\item{second_dir, }{description}
}
\value{
New sub square for second branch after taking one step
}
\description{
Moves one pixel away one second branch. Terminates if at the end of the SC.
}
