% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{get_overlap_mask}
\alias{get_overlap_mask}
\title{get_overlap_mask}
\usage{
get_overlap_mask(
  strands,
  foci_label,
  watershed_stop,
  img_orig_foci,
  watershed_radius,
  watershed_tol
)
}
\arguments{
\item{strands}{black white mask of strand channel}

\item{foci_label}{black and white mask of foci channel}

\item{watershed_stop}{Turn off default watershed method with "off"}

\item{img_orig_foci}{cropped foci channel}

\item{watershed_radius}{Radius (ext variable) in watershed method used
in foci channel. Defaults to 1 (small)}

\item{watershed_tol}{Intensity tolerance for watershed method. Defaults to 0.05.}
}
\description{
creates mask for coincident foci
}
