% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pachytene.R
\name{get_pachytene}
\alias{get_pachytene}
\title{get_pachytene}
\usage{
get_pachytene(
  img_path,
  species_num = 20,
  offset = 0.2,
  ecc_thresh = 0.85,
  area_thresh = 0.06
)
}
\arguments{
\item{img_path}{The path}

\item{file_list}{The file list}
}
\value{
Pairs of foci and SC channel crops for pachytene
}
\description{
Identifies crops in pachytene
}
