% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{keep_cells}
\alias{keep_cells}
\title{keep_cells}
\usage{
keep_cells(candidate, max_cell_area, min_cell_area)
}
\arguments{
\item{candidate}{Mask of individual cell candidates}

\item{max_cell_area, }{description}

\item{min_cell_area, }{description}
}
\value{
Mask of cell candidates which meet size criteria
}
\description{
Deletes objects in mask which are too small, large, oblong i.e. unlikely to be a cell
}
