% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{make_foci_mask}
\alias{make_foci_mask}
\title{make_foci_mask}
\usage{
make_foci_mask(
  offset_factor,
  bg,
  crowded_foci,
  img_orig_foci,
  brush_size,
  brush_sigma
)
}
\arguments{
\item{offset_factor}{Pixel value offset used in thresholding of foci channel}

\item{bg}{background value- currently just mean pixel value of whole image}

\item{crowded_foci}{TRUE or FALSE, defaults to FALSE. Set to TRUE if you
have foci > 100 or so.}

\item{img_orig_foci}{cropped foci channel}

\item{brush_size}{size of brush to smooth the foci channel. Should be small
to avoid erasing foci.}

\item{brush_sigma}{sigma for Gaussian smooth of foci channel. Should be
small to avoid erasing foci.}
}
\value{
foci mask
}
\description{
creates foci mask for foci channel crop
}
