% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{make_strand_mask}
\alias{make_strand_mask}
\title{make_strand_mask}
\usage{
make_strand_mask(offset_px, stage, img_orig, disc_size)
}
\arguments{
\item{offset_px, }{Pixel value offset used in thresholding of dna channel}

\item{stage}{meitoic stage, currently pachytene or not.}

\item{img_orig}{original strand crop}

\item{disc_size}{size of disc for local background calculation in dna channel}
}
\description{
creates strand mask for strand channel crop
}
