% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{measure_distances}
\alias{measure_distances}
\title{measure_distances}
\usage{
measure_distances(
  img_path,
  offset_px = 0.2,
  offset_factor = 2,
  brush_size = 3,
  brush_sigma = 3,
  foci_norm = 0.01,
  annotate = "off",
  offset_SC = 0.2,
  stage = "pachytene"
)
}
\arguments{
\item{img_path}{The path}

\item{file_list}{The file list}
}
\value{
Histogram of distances
}
\description{
Measure the distance between foci on a synaptonemal complex
}
