% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{remove_XY}
\alias{remove_XY}
\title{remove_XY}
\usage{
remove_XY(foci_label, foci_candidates, foci_areas)
}
\arguments{
\item{foci_label}{black and white mask of foci channel}

\item{foci_candidates}{computeFeatures data frame of foci channel}

\item{foci_areas}{the areas of the foci objects}
}
\value{
mask with XY blob removed
}
\description{
applies new row to data frame
}
