% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{threshold_SC_crop}
\alias{threshold_SC_crop}
\title{threshold_SC_crop}
\usage{
threshold_SC_crop(image, offset)
}
\arguments{
\item{image}{SC channel image}

\item{offset}{offset for an SC signal}
}
\value{
A black white mask with SCs as objects
}
\description{
Creates mask for SC channel
}
