% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{auto_crop}
\alias{auto_crop}
\title{auto_crop}
\usage{
auto_crop(
  img_path,
  crop_method = "regular",
  max_cell_area = 20000,
  min_cell_area = 7000,
  mean_pix = 0.08
)
}
\arguments{
\item{img_path}{The path}

\item{file_list}{The file list}
}
\value{
cropped SC and foci channels around single cells, regardless of stage
}
\description{
crop an image around each viable cell candidate.
}
