% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop_fast.R
\name{auto_crop_fast}
\alias{auto_crop_fast}
\title{auto_crop_fast}
\usage{
auto_crop_fast(
  img_path,
  max_cell_area = 20000,
  min_cell_area = 7000,
  mean_pix = 0.08,
  annotation = "off",
  blob_factor = 15,
  bg_blob_factor = 10,
  offset = 0.2,
  final_blob_amp = 10,
  test_amount = 0,
  brush_size_blob = 51,
  sigma_blob = 15,
  channel3_string = "DAPI",
  channel2_string = "SYCP3",
  channel1_string = "MLH3",
  file_ext = "jpeg",
  third_channel = "off"
)
}
\arguments{
\item{img_path, }{path containing image data to analyse}

\item{max_cell_area, }{Maximum pixel area of a cell candidate}

\item{min_cell_area, }{Minimum pixel area of a cell candidate}

\item{mean_pix, }{Mean pixel intensity of cell crop (in SYCP3 channel) for normalisation}

\item{annotation, }{Choice to output pipeline choices (recommended to knit)}

\item{blob_factor, }{Contrast factor to multiply original image by before smoothing/smudging}

\item{bg_blob_factor, }{Contrast factor to multiply original image by to take background. Used prior to thresholding.}

\item{offset, }{Pixel value offset from bg_blob_factor. Used in thresholding to make blob mask.}

\item{final_blob_amp, }{Contrast factor to multiply smoothed/smudged image. Used in thresholding to make blob mask.}

\item{test_amount, }{Optional number of first N images you want to run function on. For troubleshooting/testing/variable calibration purposes.}

\item{brush_size_blob, }{Brush size for smudging the dna channel to make blobs}

\item{sigma_blob, }{Sigma in Gaussian brush for smudging the dna channel to make blobs}

\item{channel3_string}{Defaults to DAPI (if third channel is on)}

\item{channel2_string}{String appended to the files showing the channel illuminating synaptonemal complexes. Defaults to SYCP3}

\item{channel1_string}{String appended to the files showing the channel illuminating foci. Defaults to MLH3}

\item{file_ext}{file extension of your images e.g. tiff jpeg or png.}

\item{third_channel}{Whether there is a third channel e.g. DAPI stain. On or off. Defaults to off.}
}
\value{
cropped SC and foci channels around single cells, regardless of stage
}
\description{
crop an image around each viable cell candidate.
}
