% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{count_foci}
\alias{count_foci}
\title{count_foci}
\usage{
count_foci(
  img_path,
  stage = "none",
  offset_px = 0.2,
  offset_factor = 2,
  brush_size = 3,
  brush_sigma = 3,
  foci_norm = 0.01,
  annotation = "off",
  channel2_string = "SYCP3",
  channel1_string = "MLH3",
  file_ext = "jpeg",
  KO_str = "--",
  WT_str = "++",
  KO_out = "-/-",
  WT_out = "+/+",
  watershed_stop = "off",
  watershed_radius = 1,
  watershed_tol = 0.05,
  crowded_foci = TRUE
)
}
\arguments{
\item{img_path, }{path containing image data to analyse}

\item{stage, }{meiosis stage of interest. Currently count_foci determines this with thresholding/ object properties in the dna channel. But will be classified using ML model in future versions.}

\item{offset_px, }{Pixel value offset used in thresholding of dna channel}

\item{offset_factor, }{Pixel value offset used in thresholding of foci channel}

\item{brush_size, }{size of brush to smooth the foci channel. Should be small to avoid erasing foci.}

\item{brush_sigma, }{sigma for Gaussian smooth of foci channel. Should be small to avoid erasing foci.}

\item{foci_norm, }{Mean intensity to normalise all foci channels to.}

\item{annotation, }{Choice to output pipeline choices (recommended to knit)}

\item{channel2_string}{String appended to the files showing the channel illuminating synaptonemal complexes. Defaults to SYCP3}

\item{channel1_string}{String appended to the files showing the channel illuminating foci. Defaults to MLH3}

\item{file_ext}{file extension of your images e.g. tiff jpeg or png.}

\item{KO_str}{string in filename corresponding to knockout genotype. Defaults to --.}

\item{WT_str}{string in filename corresponding to wildtype genotype. Defaults to ++.}

\item{KO_out}{string in output csv in genotype column, for knockout. Defaults to -/-.}

\item{WT_out}{string in output csv in genotype column, for knockout. Defaults to +/+.}

\item{watershed_stop}{Turn off default watershed method with "off"}

\item{watershed_radius}{Radius (ext variable) in watershed method used in foci channel. Defaults to 1 (small)}

\item{watershed_tol}{Intensity tolerance for watershed method. Defaults to 0.05.}

\item{crowded_foci}{TRUE or FALSE, defaults to FALSE. Set to TRUE if you have foci > 100 or so.}
}
\value{
foci count per cell
}
\description{
Count coincident foci in cropped SC and foci channel per cell
}
