% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{count_foci}
\alias{count_foci}
\title{count_foci}
\usage{
count_foci(file_list, img_path)
}
\arguments{
\item{file_list}{list of files}
}
\value{
foci count per cell
}
\description{
Count coincident foci in cropped SC and foci channel per cell
}
