% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{crop_single_object}
\alias{crop_single_object}
\title{crop_single_object}
\usage{
crop_single_object(
  retained,
  OOI_final,
  counter_final,
  img_orig,
  img_orig_foci,
  img_orig_DAPI,
  file_dna,
  file_foci,
  file_DAPI,
  cell_count,
  mean_pix,
  annotation,
  file_base,
  img_path
)
}
\arguments{
\item{retained}{Mask of cell candidates which meet size criteria. After smoothing/smudging and thresholding.}

\item{OOI_final, }{Objects of interest count. Total number of cell candidates in retained.}

\item{counter_final, }{Counter for single cell we are focussing on. Remove all other cells where counter_single not equal to counter_final.}

\item{img_orig, }{description}

\item{img_orig_foci, }{description}

\item{img_orig_DAPI, }{description}

\item{file_dna, }{filename of dna channel image}

\item{file_foci, }{filename of foci channel image}

\item{file_DAPI, }{filename of DAPI channel image}

\item{cell_count, }{counter for successful crops around cells}

\item{mean_pix, }{Mean pixel intensity of cell crop (in SYCP3 channel) for normalisation}

\item{annotation, }{Choice to output pipeline choices (recommended to knit)}

\item{file_base, }{filename base common to all three channels i.e. without -MLH3.jpeg etc.}

\item{img_path, }{path containing image data to analyse}
}
\value{
Crops around all candidates in both channels
}
\description{
Creates mask for every individual cell candidate in mask
}
