% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{crop_single_object}
\alias{crop_single_object}
\title{crop_single_object}
\usage{
crop_single_object(
  retained,
  OOI_final,
  counter_final,
  img_orig,
  img_orig_foci,
  img_orig_DAPI,
  file_dna,
  file_foci,
  file_DAPI,
  cell_count,
  mean_pix
)
}
\arguments{
\item{retained}{Mask of cell candidates which meet size criteria}
}
\value{
Crops aroudn all candidates in both channels
}
\description{
Creates mask for every individual cell candidate in mask
}
