% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{first_shot_out}
\alias{first_shot_out}
\title{first_shot_out}
\usage{
first_shot_out(
  chosen_dir,
  ix1,
  ix2,
  iy1,
  iy2,
  distance_strand,
  distance_strand_2
)
}
\arguments{
\item{chosen_dir}{string: brightest direction of a line passing through starting point}

\item{ix1, }{starting point x}

\item{ix2, }{starting point x}

\item{iy1, }{starting point y}

\item{iy2, }{starting point y}

\item{distance_strand, }{distance along first branch (zero).}

\item{distance_strand_2, }{distance along second branch (zero).}
}
\value{
New sub square for first and second branch
}
\description{
Moves one pixel away from the starting point
}
