% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop.R
\name{get_blobs}
\alias{get_blobs}
\title{get_blobs}
\usage{
get_blobs(img_orig, blob_factor, bg_blob_factor, offset, final_blob_amp)
}
\arguments{
\item{img_orig}{Original image}

\item{blob_factor, }{description}

\item{bg_blob_factor, }{description}

\item{offset, }{description}

\item{final_blob_amp, }{description}
}
\value{
Mask with cell candidates
}
\description{
Makes mask of all objects bright enough
}
