% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(
  strands,
  num_strands,
  new_img,
  foci_label,
  dimensionless_dist,
  SC_lengths,
  foci_count_strand,
  strand_iter,
  file,
  annotate
)
}
\arguments{
\item{strands}{A black white mask with SCs as objects}

\item{num_strands}{offset for an SC signal}

\item{new_img}{hello}

\item{foci_label}{A black white mask with foci as objects}

\item{dimensionless_dist}{hello}

\item{SC_lengths}{hello}

\item{foci_count_strand}{hello}

\item{strand_iter}{hello}
}
\value{
A list of distances
}
\description{
Creates mask for SC channel
}
