% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{get_distance_between_two}
\alias{get_distance_between_two}
\title{get_distance_between_two}
\usage{
get_distance_between_two(
  distance_strand,
  distance_strand_2,
  per_strand,
  foci_label,
  walkers,
  noise_gone,
  start_x,
  start_y,
  start_x2,
  start_y2,
  start_dir,
  cx,
  cy,
  mean_x,
  mean_y,
  strand_iter,
  file,
  annotate
)
}
\arguments{
\item{distance_strand, }{description}

\item{distance_strand_2, }{description}

\item{per_strand, }{description}

\item{foci_label, }{description}

\item{walkers, }{description}

\item{noise_gone, }{description}

\item{start_x, }{description}

\item{start_y, }{description}

\item{start_x2, }{description}

\item{start_y2, }{description}

\item{start_dir, }{description}

\item{cx, }{description}

\item{cy, }{description}

\item{mean_x, }{description}

\item{mean_y, }{description}

\item{strand_iter, }{description}

\item{file, }{description}

\item{annotate, }{description}
}
\value{
List of fractional distances between foci for all SCs with two. Optional: total distances of SCs. Optional: images of all resulting traces/ foci locations.
}
\description{
Calculates the pixel distance
}
