% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances_general.R
\name{get_next_first_dir}
\alias{get_next_first_dir}
\title{get_next_first_dir}
\usage{
get_next_first_dir(
  new_square_1,
  ix1,
  iy1,
  dir_1,
  window,
  chosen_dir,
  distance_strand,
  first_dir,
  SC_intens_stop
)
}
\arguments{
\item{new_square_1, }{the subsquare that contains the x y location of
first branch walker in the middle.}

\item{ix1, }{current x position of first branch walker along SC}

\item{iy1, }{current y position of first branch walker along SC}

\item{dir_1, }{The direction (choice of 8) of the first branch step}

\item{window, }{number of pixels ahead that the intensity gradient is
computed with}

\item{chosen_dir}{The brightest direction (choice of 4) of the previous step}

\item{distance_strand, }{current distance along the first branch}

\item{first_dir, }{zero while still measuring along first branch, one when
first branch terminates and counting stops.}

\item{SC_intens_stop}{mean intensity of a forward moving branch to
terminate measuring}
}
\value{
New sub square for first branch after taking one step
}
\description{
Moves one pixel away one first branch
}
