% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances_general.R
\name{get_next_second_dir}
\alias{get_next_second_dir}
\title{get_next_second_dir}
\usage{
get_next_second_dir(
  new_square_2,
  ix2,
  iy2,
  dir_2,
  window,
  chosen_dir,
  distance_strand_2,
  second_dir,
  SC_intens_stop
)
}
\arguments{
\item{new_square_2, }{the subsquare that contains the x y location of second branch walker in the middle.}

\item{ix2, }{current x position of second branch walker along SC}

\item{iy2, }{current y position of second branch walker along SC}

\item{dir_2, }{The direction (choice of 8) of the second (opposite to first) branch step}

\item{window, }{number of pixels ahead that the intensity gradient is computed with}

\item{chosen_dir}{The brightest direction (choice of 4) of the previous step}

\item{distance_strand_2, }{current distance along the second branch}

\item{second_dir, }{zero while still measuring along first branch, one when first branch terminates and counting stops.}

\item{SC_intens_stop}{mean intensity of a forward moving branch to terminate measuring}
}
\value{
New sub square for second branch after taking one step
}
\description{
Moves one pixel away one second branch. Terminates if at the end of the SC.
}
