% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pachytene.R
\name{get_pachytene}
\alias{get_pachytene}
\title{get_pachytene}
\usage{
get_pachytene(file_list, img_path)
}
\arguments{
\item{file_list}{The file list}

\item{img_path}{The path}
}
\value{
Pairs of foci and SC channel crops for pachytene
}
\description{
Identifies crops in pachytene
}
