% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pachytene.R
\name{get_pachytene}
\alias{get_pachytene}
\title{get_pachytene}
\usage{
get_pachytene(
  img_path,
  species_num = 20,
  offset = 0.2,
  ecc_thresh = 0.85,
  area_thresh = 0.06,
  annotation = "off"
)
}
\arguments{
\item{img_path, }{path containing image data to analyse}

\item{species_num, }{number of chromosomes in the species}

\item{offset, }{Pixel value offset used in therholding for the dna (SYCP3) channel}

\item{ecc_thresh, }{The minimum average eccentricity of all objects in mask determined by computefeatures, for a cell to be pachytene.}

\item{area_thresh, }{The minimum ratio of pixels included in mask to total, for a cell to be classified as pachytene.}
}
\value{
Pairs of foci and SC channel crops for pachytene
}
\description{
Identifies crops in pachytene
}
