% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_crop_fast.R
\name{keep_cells}
\alias{keep_cells}
\title{keep_cells}
\usage{
keep_cells(
  candidate,
  max_cell_area,
  min_cell_area,
  cell_aspect_ratio,
  crowded_cells,
  annotation
)
}
\arguments{
\item{candidate}{Mask of individual cell candidates}

\item{max_cell_area, }{Maximum pixel area of a cell candidate}

\item{min_cell_area, }{Minimum pixel area of a cell candidate}

\item{cell_aspect_ratio}{Maximum aspect ratio of blob to be defined as a cell}

\item{crowded_cells}{TRUE or FALSE, defaults to FALSE. Set to TRUE if you}

\item{annotation, }{Choice to output pipeline choices (recommended to knit)
have many cells in a frame that almost touch}
}
\value{
Mask of cell candidates which meet size criteria
}
\description{
Deletes objects in mask which are too small, large, oblong
i.e. unlikely to be a cell
}
