% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_foci.R
\name{make_strand_mask}
\alias{make_strand_mask}
\title{make_strand_mask}
\usage{
make_strand_mask(
  offset_px,
  stage,
  img_orig,
  disc_size,
  brush_size,
  brush_sigma
)
}
\arguments{
\item{offset_px, }{Pixel value offset used in thresholding of synaptonemal complex channel}

\item{stage, }{meiosis stage of interest. Currently count_foci determines
this with thresholding/ object properties in the synaptonemal complex channel
by previosly calling the get_pachytene function.
Note that if using this option, the count_foci function requires that the
input directory contains a folder called “pachytene” with the crops in it.}

\item{img_orig}{original strand crop}

\item{disc_size}{size of disc for local background calculation in synaptonemal complex channel}

\item{brush_size, }{size of brush to smooth the foci channel. Should be small
to avoid erasing foci.}

\item{brush_sigma, }{sigma for Gaussian smooth of foci channel. Should be
small to avoid erasing foci.}
}
\value{
strand mask
}
\description{
creates strand mask for strand channel crop
}
