% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{measure_distances}
\alias{measure_distances}
\title{measure_distances}
\usage{
measure_distances(file_list, img_path)
}
\arguments{
\item{file_list}{The file list}

\item{img_path}{The path}
}
\value{
Histogram of distances
}
\description{
Measure the distance between foci on a synaptonemal complex
}
