% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_distances.R
\name{threshold_foci_crop}
\alias{threshold_foci_crop}
\title{threshold_foci_crop}
\usage{
threshold_foci_crop(image, offset_factor, brush_size, brush_sigma)
}
\arguments{
\item{image}{foci channel image}

\item{offset_factor}{offset for a foci signal}

\item{brush_size, }{description}

\item{brush_sigma, }{description}
}
\value{
A black white mask with foci as objects
}
\description{
Creates mask for foci channel
}
